python_ml_service/pyproject.toml
[tool.poetry]
name = "pest-detection-ml-service"
version = "1.0.0"
description = "Production-ready ML microservice for pest detection using few-shot learning"
authors = ["AllergyConnectAI Team"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
torch = "^2.5.0"
torchvision = "^0.20.0"
timm = "^1.0.9"
pillow = "^11.0.0"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.12"
httpx = "^0.28.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
mypy = "^1.13.0"
ruff = "^0.7.4"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.training.dependencies]
tensorboard = "^2.18.0"
tqdm = "^4.67.0"
matplotlib = "^3.9.2"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
